Add proof-of-concept implementation of \class
A naive way of attaching \class as an override to a grob,
plus an engraver that reads and acts upon certain classes.

Classes must not be defined directly within the engraver, but the engraver
will have to look them up somewhere (where some \defineClass may
store them).

Is it possible to define the classes in LilyPond language? It seems not,
the engraver may have to directly access grob properties.